pixel_map: render just P for a known match
Now that we know what the pixels look like and what some characters look
like, render a known manually crafted one and output its bit patterns
for grepping them in the binary in the hope to find a font bitmap.

One opportunistic way to look for it would be to clock in the firmware
binaries as-is, with varying offsets, and to observe when a good pattern
emerges, if any.