pixel_map: Introduce font explorer
The fonts were found to be located in the zel09101 ROM image.

As the glyph P acted as the origin of this information, add
experimentation to render things offset from the bit address of that
glyph. There seems to be exactly 256 glyphs.

The display was held upside down when decoding the segments so the
original pixel map needs refactoring in the future.