pixel_map: cleanup font, add text rendering
The font bitmap in zel09101 starts at 0x400 (1KB) and is 256 glyphs
long, 15 bytes each. Refactor the glyph drawing code to be based on
these numbers and make the demo loop use it.

Add a text rendering function to start with. The bitmap is arranged in
very much ascii order.