pixel_map: support multiple panels
Include panel offset in the segment address math. The panels are
addressed in the current "API" based on the digit. Digits 0-3 go to the
first panel, digits 4-7 go to the second panel and so forth.

In the wire, the data for the last panel in the chain naturally goes in
first, as it's a shift register.