pixel_gfx: initial messing around
Introduce a graphical emulator for the physical displays. For each
segment, define a polygon and draw it at the right screen offset based
on the corresponding square region. The polygons are built with a simple
stack machine because that's fun.

This is not ready yet. This is a separate patch to keep the next one
from becoming too big. Maybe. And to document the development at early
stages.