pixel proxy: use HW SPI
Hardware SPI gives a bit more speed than bitbanging the bus. Note that
this requires a modification to the wiring between arduino and the lcd.

Rather than SPI.transfer() that busyloops until done, write to SPDR
directly and only wait for the bit fifo to be empty before sending new
bits, not after. Also wait before latching the display data upon
completion of a frame, but do that after Serial.write() to hide some
latency there as well.