pixel proxy: autodetect panel count
Feed in a bit pattern and see when it comes out to detect the shift
register length. Try this with three different values just in case to
ensure signal integrity. This needs some more HW modifications: the DO
pin of the last module in the chain must be connected to the SPI in pin.

Previously the display would latch once every panel boundary. With the
latch strobed once all modules have received data, the next panels in
line no longer shadow the first (or last, depending on how you look at
it) for that split second that was noticeable in fast animations.